+++ Samples/JotFormTest/Classes/Controller/CreateSubmissionViewController.m
@@ -57,8 +57,7 @@
         [SVProgressHUD dismiss];
         
         if ( result != nil ) {
-            
-            int responseCode = [[result objectForKey:@"responseCode"] integerValue];
+            NSInteger responseCode = [[result objectForKey:@"responseCode"] integerValue];
             
             if ( responseCode == 200 || responseCode == 206 ) {
                 
@@ -73,7 +72,7 @@
         [SVProgressHUD dismiss];
         
         if (error != nil) {
-            int responseCode = [[error objectForKey:@"responseCode"] integerValue];
+           NSInteger responseCode = [[error objectForKey:@"responseCode"] integerValue];
             
             if ( responseCode == 401 ) {
