+++ Samples/JotFormTest/Classes/Controller/RegisterUserViewController.m
@@ -81,8 +81,7 @@
         [SVProgressHUD dismiss];
         
         if ( result != nil ) {
-            
-            int responseCode = [[result objectForKey:@"responseCode"] integerValue];
+            NSInteger responseCode = [[result objectForKey:@"responseCode"] integerValue];
             
             if ( responseCode == 200 || responseCode == 206 ) {
